fix: move pages to pages-v2
