+++ apps/gitness/src/pages-v2/repo/repo-commit-details-diff.tsx
@@ -15,9 +15,9 @@ import { CommitDiff, CommitSidebar } from '@harnessio/ui/views'
 import Explorer from '../../components-v2/FileExplorer'
 import { useGetRepoRef } from '../../framework/hooks/useGetRepoPath'
 import useCodePathDetails from '../../hooks/useCodePathDetails'
-import { parseSpecificDiff } from '../../pages/pull-request/diff-utils'
 import { PathParams } from '../../RouteDefinitions'
 import { normalizeGitRef } from '../../utils/git-utils'
+import { parseSpecificDiff } from '../pull-request/diff-utils'
 import { changedFileId, DIFF2HTML_CONFIG, normalizeGitFilePath } from '../pull-request/pull-request-utils'
 import { useCommitDetailsStore } from './stores/commit-details-store'
