--- apps/gitness/src/pages/pull-request/hooks/useDataFilters.ts
@@ -1,19 +1,0 @@
-import { useMemo } from 'react'
-
-import { orderSortDate } from '../types/types'
-
-export function useDateFilters() {
-  return useMemo(
-    () => [
-      {
-        label: 'First added',
-        value: orderSortDate.ASC
-      },
-      {
-        label: 'Last added',
-        value: orderSortDate.DESC
-      }
-    ],
-    []
-  )
-}
